Add AgentHandler class for save/load/eval of agents. Remove AgentEvaluator. Adapt train_eval and eval scripts.
